cicd: updated collections in content.config.ts, created a separate schema file.
