+++ src/pages/release-notes/[...slug].astro
@@ -11,7 +11,7 @@ import { stripMarkdownSlug } from "~/utils/slugUtils";
 
 // New way of getting toc
 import { getTableOfContentsFromCollection } from "~/utils/getTableOfContentsFromCollection";
-const collectionName = "release-notes";
+const collectionName = `superoffice-docs/release-notes`;
 const allTocEntries = await getCollection("tocExternal");
 const tocEntries = allTocEntries.filter((entry) =>
   entry.id.startsWith(collectionName)
@@ -33,7 +33,10 @@ export async function getStaticPaths() {
   return docEntries.map((entry, index) => {
 
     //Remove .md extension AND "src/content/release-notes" from filepath
-    const generatedSlug = stripMarkdownSlug(entry.filePath!, collectionName, false);
+    const generatedSlug = entry.filePath
+      ?.replace("external-content/superoffice-docs/release-notes", "")
+      .replace(".md", "");
+    // const generatedSlug = stripMarkdownSlug(entry.filePath!, collectionName, false);
 
     return {
       params: { slug: `${generatedSlug}` },
